% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requisicao.R
\name{aux_retorna_endpoint}
\alias{aux_retorna_endpoint}
\title{Retorna o endpoint correspondente ao tribunal informado}
\usage{
aux_retorna_endpoint(tribunal)
}
\arguments{
\item{tribunal}{Um caractere que representa o código ou sigla do tribunal.}
}
\value{
Caractere com a URL do endpoint correspondente ao tribunal informado.
}
\description{
Esta função auxiliar é destinada a identificar e retornar a URL do endpoint
específico de um tribunal para consulta na API pública do Datajud.
}
\examples{
aux_retorna_endpoint("TST")
aux_retorna_endpoint("trf01")

}
