% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_dados.R
\name{datajud_desaninhar_assuntos}
\alias{datajud_desaninhar_assuntos}
\title{Desaninha os assuntos estruturados dos processos}
\usage{
datajud_desaninhar_assuntos(dados)
}
\arguments{
\item{dados}{Resultado de \code{datajud_ler_processo()}.}
}
\value{
Tibble com uma linha por processo e assunto.
}
\description{
Desaninha os assuntos estruturados dos processos
}
