% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesquisa_assuntos.R
\name{datajud_pesquisar_classe_orgao}
\alias{datajud_pesquisar_classe_orgao}
\title{Pesquisa processos no Datajud por Classe e/ou Órgão Julgador}
\usage{
datajud_pesquisar_classe_orgao(
  tribunal,
  cliente,
  classe_codigo = NULL,
  orgao_codigo = NULL,
  size = 100
)
}
\arguments{
\item{tribunal}{Identificador do tribunal a ser consultado.}

\item{cliente}{Objeto criado por \code{datajud_cliente()}.}

\item{classe_codigo}{Vetor opcional de códigos de classe para filtrar os processos.}

\item{orgao_codigo}{Vetor opcional de códigos de órgão julgador para filtrar os processos.}

\item{size}{Tamanho máximo da amostra de resultados a ser retornada, com um valor padrão de 100. O tamanho máximo permitido é 10000.}
}
\value{
Uma lista com os processos encontrados pela consulta.
}
\description{
Esta função realiza uma pesquisa no Datajud, permitindo filtrar processos por código de classe e/ou por código de órgão julgador.
É possível especificar um tamanho máximo para a amostra de resultados retornados.
}
\examples{
\dontrun{
# Pesquisar processos no TJSP por classe de assunto 1116 e tamanho da amostra de 100
datajud_pesquisar_classe_orgao(tribunal = "TJSP", classe_codigo = c(1116), size = 100)

# Pesquisar processos no TJMG por órgão julgador 13597 com o tamanho padrão da amostra
datajud_pesquisar_classe_orgao(tribunal = "TJMG", orgao_codigo = c(13597))

# Pesquisar processos no TJRJ por classe de assunto e órgão julgador especificados
datajud_pesquisar_classe_orgao(tribunal = "TJRJ", classe_codigo = c(1116), orgao_codigo = c(13597), size = 500)
}
}
