% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{obter_chave_publica_cnj}
\alias{obter_chave_publica_cnj}
\title{Obtém a chave pública vigente na Wiki do CNJ}
\usage{
obter_chave_publica_cnj(
  url = "https://datajud-wiki.cnj.jus.br/api-publica/acesso/"
)
}
\arguments{
\item{url}{URL da página oficial de acesso da API.}
}
\value{
A chave pública como caractere de comprimento um.
}
\description{
Obtém a chave pública vigente na Wiki do CNJ
}
