% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_request.R
\name{parse_response}
\alias{parse_response}
\title{Parse DataJud API Response}
\usage{
parse_response(response)
}
\arguments{
\item{response}{List containing the raw API response.}
}
\value{
A tibble with the parsed results.
}
\description{
Parses the raw API response into a more user-friendly data frame.
}
\keyword{internal}
