% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultado.R
\name{print.datajud_resultado}
\alias{print.datajud_resultado}
\title{Imprimir um resultado de pesquisa do Datajud}
\usage{
\method{print}{datajud_resultado}(x, ...)
}
\arguments{
\item{x}{Objeto \code{datajud_resultado}.}

\item{...}{Argumentos adicionais, atualmente ignorados.}
}
\value{
O próprio objeto, invisivelmente.
}
\description{
Imprimir um resultado de pesquisa do Datajud
}
