% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribunals.R
\name{validate_tribunal}
\alias{validate_tribunal}
\title{Validate Tribunal Alias}
\usage{
validate_tribunal(tribunal)
}
\arguments{
\item{tribunal}{Character string with the tribunal alias.}
}
\value{
TRUE if valid, FALSE otherwise.
}
\description{
Checks if a tribunal alias is valid.
}
\keyword{internal}
