% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requisicao.R
\name{datajud_consultar_processo}
\alias{datajud_consultar_processo}
\title{Title}
\usage{
datajud_consultar_processo(processo, tribunal = NA, sleep = 0.1)
}
\arguments{
\item{tribunal}{}
}
\description{
Title
}
