% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_dados.R
\name{datajud_ler_movimentacoes}
\alias{datajud_ler_movimentacoes}
\title{Lê as movimentações de processos retornadas pelo Datajud}
\usage{
datajud_ler_movimentacoes(base)
}
\arguments{
\item{base}{Lista contendo os dados dos processos retornados pela API.}
}
\value{
Imprime e retorna um data frame consolidado com as movimentações de todos os
processos fornecidos. Cada linha representa uma movimentação específica, incluindo
metadados relevantes para análises subsequentes.
}
\description{
Esta função é utilizada para extrair e processar as movimentações dos processos
judiciais obtidos a partir de uma pesquisa no Datajud. Ela opera sobre uma lista
de processos fornecida diretamente.
É ideal para análises detalhadas das etapas processuais e suas características.
}
\examples{
\dontrun{
datajud_ler_movimentacoes(resposta)
}
}
