% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_dados.R
\name{datajud_ler_processo}
\alias{datajud_ler_processo}
\title{Lê os dados de processos retornados pelo Datajud}
\usage{
datajud_ler_processo(base)
}
\arguments{
\item{base}{Lista de respostas de processos retornadas pela API.}
}
\value{
Um tibble contendo os metadados dos processos.
}
\description{
Após realizar uma pesquisa de processos com \code{datajud_pesquisar_classe_orgao},
esta função permite ler e manipular os dados dos processos retornados.
}
\examples{
\dontrun{
cliente <- datajud_cliente()
resposta <- datajud_consultar_processo(
  "0000001-89.2020.8.26.0000",
  tribunal = "TJSP",
  cliente = cliente
)
datajud_ler_processo(resposta)
}
}
