% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explora_assuntos.R
\name{datajud_pesquisar_classe_orgao}
\alias{datajud_pesquisar_classe_orgao}
\title{Title}
\usage{
datajud_pesquisar_classe_orgao(
  tribunal = NA,
  lista_classe = NULL,
  lista_orgao = NULL,
  size = 100
)
}
\arguments{
\item{size}{}
}
\description{
Title
}
